ed_rag/pyproject.toml
[tool.poetry]
name = "ed-rag"
version = "0.1.0"
description = ""
authors = ["Durgesh Nandan <durgnan45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.6"
langchain-core = "^0.2.11"
langchain-community = "^0.2.6"
langchain-experimental = "^0.0.62"
langgraph = "^0.1.5"
langserve = {extras = ["all"], version = "^0.2.2"}
langchain-cli = "^0.0.25"
langsmith = "^0.1.83"
pandas = "^2.2.2"
langchain-openai = "^0.1.16"
beautifulsoup4 = "^4.12.3"
pypdf = "^4.2.0"
faiss-cpu = "^1.8.0.post1"
sentence-transformers = "^3.0.1"
langchain-chroma = "^0.1.2"
langchainhub = "^0.1.20"
langchain-huggingface = "^0.0.3"
flashrank = "^0.2.6"
flask = "^3.0.3"
flask-cors = "^4.0.1"
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"
flask-restful = "^0.3.10"
mongoengine = "^0.28.2"
flask-socketio = "^5.3.6"
redis = "^5.0.8"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
